requirements.txt
python-telegram-bot
python-dotenv
cerebras_cloud_sdk
colorama
psutil
